+++ controllers/adminDataFunction/login.js
@@ -1,5 +1,5 @@
 const { StatusCodes } = require("http-status-codes");
-const adminLoginService = require("../../service/adminService/adminLogin");
+const adminLoginService = require("../../service/adminService/adminLogin.service");
 const passport = require("passport");
 
 exports.loginPage = (req, res) => {
